cache grep

+++ pkgs/grep/fetch.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 rm -rf "$_DESTDIR" "$_SRCDIR"
-printf "Downloading source...\n"
-curl -L -# -o src.tar.xz https://ftpmirror.gnu.org/gnu/grep/grep-3.12.tar.xz
+ver='3.12'
+if [ ! -f "$_DLCACHE/grep-$ver.tar.xz" ] ||
+    [ "$(sha256sum "$_DLCACHE/grep-$ver.tar.xz" | awk '{print $1}')" != "2649b27c0e90e632eadcd757be06c6e9a4f48d941de51e7c0f83ff76408a07b9" ]; then
+    printf "Downloading source...\n"
+    curl -L -# -o "$_DLCACHE/grep-$ver.tar.xz" "https://ftpmirror.gnu.org/gnu/grep/grep-$ver.tar.xz" || exit 1
+fi
 printf "Unpacking source...\n"
-tar -C "$_TMP" -xf src.tar.xz
-rm src.tar.xz
+tar -C "$_TMP" -xf "$_DLCACHE/grep-$ver.tar.xz"
 mv "$_TMP"/grep-* "$_SRCDIR"
